% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Builds a PDF report from MBE702 response survey CSV}
\usage{
build_report(infile)
}
\arguments{
\item{infile}{a \code{STRING} path to where the target .csv survey results file is located}
}
\description{
Builds a PDF report from MBE702 response survey CSV
}
