% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Checks matrix for proper formating}
\usage{
check_input(ma)
}
\arguments{
\item{ma}{input matrix.}
}
\value{
\code{TRUE} if properly formated, \code{FALSE} if not.
}
\description{
Checks matrix for proper formating
}
\examples{
check_input(ma)
}
