% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_boxplot.R
\name{report_boxplot}
\alias{report_boxplot}
\title{Generates a boxplot from MBE702 survey data (not for end user)}
\usage{
report_boxplot(dats, plot = TRUE)
}
\arguments{
\item{dats}{a list object returned from \code{process_input()}}

\item{plot}{a \code{BOOL} to determine if to plot the result. Defaults to \code{TRUE}.}
}
\value{
an invisible ggplot boxplot of the input data.
}
\description{
Generates a boxplot from MBE702 survey data (not for end user)
}
\examples{
report_boxplot(ma)
}
