% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_comments.R
\name{report_comments}
\alias{report_comments}
\title{Builds a list containing Roxygen formated strings used for generating MBE702 survey results report (not for end user)}
\usage{
report_comments(dats)
}
\arguments{
\item{dats}{a list object returned from \code{process_input()}}
}
\value{
a heatmap of responses.
}
\description{
Builds a list containing Roxygen formated strings used for generating MBE702 survey results report (not for end user)
}
