% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_heatmap.R
\name{report_heatmap}
\alias{report_heatmap}
\title{Builds a response heatmap from an MBE702 survey (not for end user)}
\usage{
report_heatmap(dats)
}
\arguments{
\item{dats}{a list object returned from \code{process_input()}}
}
\value{
a heatmap of responses.
}
\description{
Builds a response heatmap from an MBE702 survey (not for end user)
}
