% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_strengths.R
\name{report_strengths}
\alias{report_strengths}
\title{Builds a list containing Roxygen formated strings used for generating MBE702 survey results report (not for end user)}
\usage{
report_strengths(dats)
}
\arguments{
\item{dats}{a list object returned from \code{process_input()}}
}
\value{
a histogram of responses.
}
\description{
Builds a list containing Roxygen formated strings used for generating MBE702 survey results report (not for end user)
}
