% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Builds a PDF report from MBE702 response survey CSV}
\usage{
build_report(infile)
}
\arguments{
\item{path}{a \code{STRING} path to where the target .csv survey results file is located}
}
\value{
does not return, instead it generates a PDF result file
}
\description{
Builds a PDF report from MBE702 response survey CSV
}
\examples{
build_report("~/Desktop/MBE702/2019/Fall/test.csv")
}
